requirements.txt
flask
tensorflow
pillow
numpy
scikit-learn
